requirements.txt
aiohttp==3.9.3
chromadb==0.4.24
clip==1.0
gradio==4.26.0
html2text==2024.2.26
ipython==8.12.3
langchain==0.1.16
langchain_community==0.0.32
llama_index==0.10.28
llava==0.0.1.dev0
matplotlib==3.8.3
numpy==1.26.4
openai==1.17.0
pandas==2.2.2
peft==0.10.0
Pillow==10.3.0
Requests==2.31.0
sentence_transformers==2.6.1
torch==2.1.2+cu118
tqdm==4.66.2
transformers==4.39.2
